Updated position and duration calls to use long instead of int (#388)

+++ library/src/main/java/com/devbrackets/android/exomedia/core/video/exo/ExoTextureVideoView.java
@@ -91,7 +91,7 @@ public class ExoTextureVideoView extends ResizingTextureView implements VideoVie
     }
 
     @Override
-    public void seekTo(@IntRange(from = 0) int milliseconds) {
+    public void seekTo(@IntRange(from = 0) long milliseconds) {
         delegate.seekTo(milliseconds);
     }
 
@@ -121,12 +121,12 @@ public class ExoTextureVideoView extends ResizingTextureView implements VideoVie
     }
 
     @Override
-    public int getDuration() {
+    public long getDuration() {
         return delegate.getDuration();
     }
 
     @Override
-    public int getCurrentPosition() {
+    public long getCurrentPosition() {
         return delegate.getCurrentPosition();
     }
